requirements.txt
# Web Scraping
playwright==1.49.0
httpx>=0.27.0

# Data Processing
pandas>=2.2.0
pydantic>=2.5.0
beautifulsoup4>=4.12.3
lxml>=5.2.1

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9    # PostgreSQL adapter (for Supabase)
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.4
tenacity>=8.2.0

# Testing
pytest>=7.4.0

# AI / Embeddings
openai>=1.12.0
numpy>=1.26.0
tiktoken>=0.6.0  # Recommended for token counting

# 설치 후 반드시 실행:
# playwright install chromium